Registering New Users
Registering New User

Implemented the new user registration feature to add a new user object to a newly added users object global variable. Updated the cookie value being saved and also refactored the routes to pass a user object to the templates rather than the username. Also refatcored the header partials to implement the use if the user object passed.